remove: 사용하지 않는 파일 정리
