--- src/main/java/subway/dto/LineRequest.java
@@ -1,23 +1,0 @@
-package subway.dto;
-
-public class LineRequest {
-    private String name;
-    private String color;
-
-    public LineRequest() {
-    }
-
-    public LineRequest(String name, String color) {
-        this.name = name;
-        this.color = color;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-}
